Remove dependency on versionomy gem.
The versionomy gem requires a native install which would require build
dependencies to be installed, increasing the dependency chain for this
cookbook.  This commit uses Gem::Version comparison and removes the
versionomy dependencies.

The implementation is a bit messy since we need to preserve the build
slugs at the end of the version that come out of Omnitruck, but we
don't want to compare them.

This implementation also fixes a bug where :latest would not correctly
sort based on build.  Note, however, that I've kept two potentially
errant behaviors:

- "latest" will install prerelease packages,
- It is impossible to specify the version without also knowing the
 build slug.

Ultimately, the omnitruck asset endpoint should simply return the
correct package based on parameters to the API request, just like the
client endpoint does currently.