Ensure creation of `:file_cache_path`
Fix to ensure that the file cache path exists. An error occurs while testing the recipe with Vagrant, which was narrowed to the path not existing when downloading the the chef-server remote package (`package_name`). The path created has proper permissions, and relies on Chef's config hash value for the path.