Refactor for Chef Server 12
This commit refactors the cookbook for Chef Server 12. The README
covers the new scope of this cookbook. We’ve attempted to preserve
parity in platform support with that of Chef Server 12 itself.

* Remove the omnitruck API client library and related attributes
* Use packagecloud repository for packages through the
chef_server_ingredient resource
* Remove the “dev” recipe
* Remove the dependency on the git cookbook
* Remove the Vagrantfile - we use test-kitchen now
* Update the Berksfile accordingly
* Add ServerSpec tests

Fixes #46, COOK-4691: use IP address as hostname.

Fixed merge conflicts:

	.kitchen.yml
	recipes/default.rb

Added tests for the no-fqdn suite.

whitespace, be aligned correctly

chef-solo needs dependencies

In the README steps for doing the chef-solo run with this cookbook we
also need to get the dependencies in place, since we don’t have a
depsolver to do this.

also update the cloud in the kitchen

it's cool when you forget to save all changes

use supermarket for all cookbooks in the solo instructions