Fixes #74, use reconfigure action
Instead of notifying that the ingredient should reconfigure, we should
use the reconfigure action in the default recipe. It's important to
note that while this is idempotent, but it is not convergent, and will run
every time chef-client runs on the Chef Server.

Also add a recipe to create a user and org, and serverspec to test
they exist.