Added the ability to specify a local package for chef-ingredient to use for addons.
The syntax for adding addons with packagecloud will stay the same.
To add a local package you just need to specify the attribute:
node['chef-server']['addons'] = [ { 'manage' => '~/mypackage.deb' } ]

The package needs to be located on the VM so to test you will probably
want to add a synced folder to kitchen.yml.